Build the favicon design section of the request

This function helps with the creation of the 'favicon_design' section of the request. In turn, this function is fed by a number of other helper functions designed for each platform.

build_favicon_design(
  ios = ios_helper(),
  windows = windows_helper(),
  firefox_app = firefox_app_helper(),
  android_chrome = android_chrome_helper(),
  safari_pinned_tab = safari_pinned_tab_helper(),
  coast = coast_helper(),
  open_graph = open_graph_helper(),
  yandex_browser = yandex_browser_helper()
)

Arguments

ios
list; list of ios parameters. Most easily created using the ios_helper() function.

windows
list; list of ios parameters. Most easily created using the windows_helper() function.

firefox_app
list; list of ios parameters. Most easily created using the firefox_app_helper() function.

android_chrome
list; list of ios parameters. Most easily created using the android_chrome_helper() function.

safari_pinned_tab
list; list of ios parameters. Most easily created using the safari_pinned_tab_helper() function.

coast
list; list of ios parameters. Most easily created using the coast_helper() function.

open_graph
list; list of ios parameters. Most easily created using the open_graph_helper() function.

yandex_browser
list; list of ios parameters. Most easily created using the yandex_browser_helper() function.

Value

list; complete list of favicon parameters for each platform.